% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{as_chronicle}
\alias{as_chronicle}
\title{Coerce an object to a chronicle object.}
\usage{
as_chronicle(.x, .log_df = data.frame())
}
\arguments{
\item{.x}{Any object.}

\item{.log_df}{Used internally, the user does need to interact with it. Defaults to an empty data frame.}
}
\value{
Returns a chronicle object with the object as the $value.
}
\description{
Coerce an object to a chronicle object.
}
\examples{
as_chronicle(3)
}
