% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{browser_addin_binding}
\alias{browser_addin_binding}
\title{Browser binding.}
\usage{
browser_addin_binding(.inBrowser = FALSE)
}
\arguments{
\item{.inBrowser}{Run in a browser instead of a in-process popup.}
}
\description{
Browser binding.
}
