% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicler.R
\docType{package}
\name{chronicler}
\alias{chronicler}
\alias{chronicler-package}
\title{Chronicler: Binary History for R}
\description{
Meta-package that installs all dependencies needed to
enable history of binary R objects: data sets, models, plots, code.

Package-level \link{options}:
\itemize{
\item \code{chronicler.attach} (default: \code{TRUE}) find the most
up-to-date \link{commit} in the repository and continue work from
there
}
}
