% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{dictionary}
\alias{dictionary}
\alias{commit}
\title{Dictionary of terms.}
\description{
Below are definitions which conceptually organize the
repository of artifacts.

\itemize{
\item \code{commit} is a record of the state of the R session: its \code{\link[=globalenv]{globalenv()}},
the last plot, the last printout, etc.; repository keeps track of
artifacts by recording commits throught the duration of each R session
}
}
