% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\name{engine-chronicler}
\alias{engine-chronicler}
\alias{is_artifact_id}
\alias{knit_artifact}
\title{Knit artifacts.}
\usage{
is_artifact_id(x)

knit_artifact(id)
}
\arguments{
\item{x}{value to be tested.}

\item{id}{artifact identifier.}
}
\description{
Knit artifacts.
}
