% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments}
\alias{experiments}
\alias{find_experiments}
\title{Experiments.}
\usage{
find_experiments()
}
\value{
\code{find_experiments} returns a container (see \link[utilities:as_container]{utilities::as_container})
of objects of class \code{"experiment"}.
}
\description{
Experiments.

\code{show_experiments} finds all artifacts matching the definition of a
model (e.g. of class \code{"lm"}) together with their:
\itemize{
\item parametrization: literal scalars and scalar variables used in
expressions/calls leading up to the given artifact
\item sequence of commands from the original data set to that model
\item list of downstream artifacts, like predictions, model evaluations,
plots, etc.
}
}
