% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_functions.R
\name{filter2}
\alias{filter2}
\title{Filter a data frame using predicates}
\usage{
filter2(...)
}
\arguments{
\item{...}{Further arguments passed down to \code{dplyr::filter()}.}
}
\value{
The filtered data frame
}
\description{
Filter a data frame using predicates
}
\details{
This function is nothing but a wrapper around \code{dplyr::filter()}, but raises a warning
if the output is an empty data frame.
}
