% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{fmap_record}
\alias{fmap_record}
\title{Evaluate a non-chronicle function on a chronicle object.}
\usage{
fmap_record(.c, .f, ...)
}
\arguments{
\item{.c}{A chronicle object.}

\item{.f}{A non-chronicle function.}

\item{...}{Further parameters to pass to \code{.f}.}
}
\value{
Returns the result of \code{.f(.c$value)} as a new chronicle object.
}
\description{
Evaluate a non-chronicle function on a chronicle object.
}
\examples{
as_chronicle(3) |> fmap_record(sqrt)
}
