% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwards.R
\docType{import}
\name{forwards}
\alias{forwards}
\alias{reexports}
\alias{iris_model}
\alias{london_meters}
\title{API forwarded from other packages.}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{repository}{\code{\link[repository]{iris_model}}, \code{\link[repository]{london_meters}}}
}}

