% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{\%>=\%}
\alias{\%>=\%}
\title{Pipe a chronicle object to a decorated function.}
\usage{
.c \%>=\% .f
}
\arguments{
\item{.c}{A value returned by record.}

\item{.f}{A chronicle function to apply to the returning value of .c.}
}
\value{
A chronicle object.
}
\description{
Pipe a chronicle object to a decorated function.
}
\examples{
r_sqrt <- record(sqrt)
r_exp <- record(exp)
3 |> r_sqrt() \%>=\% r_exp()
}
