% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R, R/record_many.R
\name{is_chronicle}
\alias{is_chronicle}
\title{Checks whether an object is of class "chronicle"}
\usage{
is_chronicle(.x)

is_chronicle(.x)
}
\arguments{
\item{.x}{An object to test.}
}
\value{
TRUE if .x is of class "chronicle", FALSE if not.

TRUE if .x is of class "chronicle", FALSE if not.
}
\description{
Checks whether an object is of class "chronicle"

Checks whether an object is of class "chronicle"
}
