% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{make_log_df}
\alias{make_log_df}
\title{Creates the log_df element of a chronicle object.}
\usage{
make_log_df(
  ops_number = 1,
  success,
  fstring,
  args,
  res_pure,
  start = Sys.time(),
  end = Sys.time(),
  .g = (function(x) NA),
  diff_obj = NULL
)
}
\arguments{
\item{ops_number}{Tracks the number of the operation in a chain of operations.}

\item{success}{Did the operation succeed?}

\item{fstring}{The function call.}

\item{args}{The arguments of the call.}

\item{res_pure}{The result of the purely call.}

\item{start}{Starting time.}

\item{end}{Ending time.}

\item{.g}{Optional. A function to apply to the intermediary results for monitoring purposes. Defaults to returning NA.}

\item{diff_obj}{Optional. Output of the \code{diff} parameter in \code{record()}.}
}
\value{
A tibble containing the log.
}
\description{
Creates the log_df element of a chronicle object.
}
