% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{read.log}
\alias{read.log}
\title{Reads the log of a chronicle.}
\usage{
read.log(.c)
}
\arguments{
\item{.c}{A chronicle object.}
}
\value{
The log of the object.
}
\description{
Reads the log of a chronicle.
}
\examples{
\dontrun{
read.log(chronicle_object)
}
}
