% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_functions.R
\name{select2}
\alias{select2}
\title{Select a data frame using predicates}
\usage{
select2(...)
}
\arguments{
\item{...}{Further arguments passed down to \code{dplyr::select()}.}
}
\value{
The selected data frame
}
\description{
Select a data frame using predicates
}
\details{
This function is nothing but a wrapper around \code{dplyr::select()}, but raises a warning
if the output is an empty data frame.
}
