% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R, R/ui.R
\docType{data}
\name{chronicler_state}
\alias{chronicler_state}
\alias{prepare_session}
\alias{init_ui}
\title{Internal state object, used by the \code{ui} package.}
\format{An object of class \code{environment} of length 2.}
\usage{
chronicler_state

prepare_session(state, path)

init_ui(state)
}
\arguments{
\item{state}{the global session state object.}
}
\description{
\code{prepare_session} runs the logic of attaching to
a repository without any user interactions. In case of an error, it
informs the user how to start the interactive wizard.
}
\keyword{datasets}
