% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{write_chronicle}
\alias{write_chronicle}
\title{Save 'chronicle' objects where the value is a data.frame to disk}
\usage{
write_chronicle(.c, path, row.names = FALSE, sep = ",", ...)
}
\arguments{
\item{.c}{A chronicle object.}

\item{path}{Path to the file. The extension (one of '.csv' or '.xlsx') determines the format of the file.}

\item{row.names}{Should row names be saved? Defaults to FALSE.}

\item{sep}{The separator for '.csv' files. Defaults to ",".}

\item{...}{Further arguments passed down to \code{read.table()} or \code{openxlsx::write.xlsx()}.}
}
\value{
No return value, called for side effects (writing data to disk).
}
\description{
Save 'chronicle' objects where the value is a data.frame to disk
}
\details{
'chronicle' objects where the "$value" is a data.frame can be saved with
their log in either the '.csv' or '.xlsx' format to disk. In case
the data is saved in the '.csv' format, the first \emph{n} lines will contain
the log. The number of lines that should be skipped to read in the data are
indicated. In case the data gets saved in the '.xlsx' format, the data
is written on the first sheet, and the log on the second sheet. \code{read.table()}
with \code{row.names = FALSE} and \code{sep = ","} is used to save to the '.csv' format,
and \code{openxlsx::write.xlsx()} is used to save to the '.xlsx' format.
}
