% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{bind_record}
\alias{bind_record}
\title{Evaluate a decorated function}
\usage{
bind_record(.c, .f, ...)
}
\arguments{
\item{.c}{A chronicle object (a list of two elements)}

\item{.f}{A chronicle function to apply to the returning value of .c}

\item{...}{Further parameters to pass to .f}
}
\value{
A list with elements .f(.c$value) and concatenated logs.
}
\description{
Evaluate a decorated function
}
\examples{
r_sqrt <- record(sqrt)
r_exp <- record(exp)
3 |> r_sqrt() |> bind_record(r_exp)
}
