% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{chronicler-ui}
\alias{chronicler-ui}
\alias{artifacts}
\alias{history}
\alias{identify_artifact}
\alias{explain}
\title{Chronicler user interface objects.}
\format{An object of class \code{NULL} of length 0.}
\usage{
artifacts

history

identify_artifact(x)

explain(x)
}
\arguments{
\item{x}{R object, file path or an artifact.}
}
\description{
Chronicler user interface objects.

\code{artifacts} is the entry point for artifact search. It exposes an
interactive query mechanism with tab-completion via the \code{$} (dollar)
operator.

\code{history} gives access to the history of commands stored in the
repositoy along artifacts.

\code{identify} takes an object or a file path as a parameter and returns
artifacts matching that object or contents of the file. A number of
file types are supported: plots, text data, serialized R objects.

\code{explain} provides additional context for an artifact.
}
\examples{
\dontrun{
# show summary of all artifacts in repository
artifacts

# show all R sessions
artifacts$session

# show artifacts from a given session
artifacts$session$afc61b89

# show all artifacts with a given name
artifacts$name$input
artifacts$input

# retrieve a given artifact
artifacts$id$c643b1a$value
}
\dontrun{
# look for an object
identify(iris)

# look for a plot
identify('./plot.png')
}
}
\keyword{datasets}
