% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R, R/ui.R
\docType{data}
\name{attach_wizard}
\alias{attach_wizard}
\alias{artifacts}
\title{Search for artifacts.}
\format{An object of class \code{NULL} of length 0.}
\usage{
attach_wizard()

artifacts
}
\description{
\code{chronicler_attach} extends the default, on-load logic
with interactions with the user whenever a decision cannot be made
automatically and thus results in an error on package load.

An entry point for artifact search. Exposes an interactive query
mechanism with tab-completion via the \code{$} (dollar) operator.
}
\keyword{datasets}
