% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{ggplot_fun}
\alias{ggplot_fun}
\title{Helper function for record_ggplot()}
\usage{
ggplot_fun(ggplot_expression)
}
\arguments{
\item{ggplot_expression}{An unevaluated ggplot expression.}
}
\value{
A ggplot object.
}
\description{
ggplot_fun() takes a ggplot_expression (an unevaluated ggplot expression) as input
and does the following steps:
\enumerate{
\item Evaluates the ggplot_expression using rlang::eval_tidy
\item Builds the ggplot object using ggplot2::ggplot_build
\item Returns the ggplot object
}
}
\examples{
\dontrun{
# This function is not meant to be used directly by the user.
# Instead, use record_ggplot function.
}
}
