% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{print.chronicle}
\alias{print.chronicle}
\title{Print method for chronicle objects.}
\usage{
\method{print}{chronicle}(x, ...)
}
\arguments{
\item{x}{A chronicle object.}

\item{...}{Unused.}
}
\description{
Print method for chronicle objects.
}
