% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{record_ggplot}
\alias{record_ggplot}
\title{Record ggplot}
\usage{
record_ggplot(ggplot_expression, strict = 2)
}
\arguments{
\item{ggplot_expression}{A ggplot expression.}

\item{strict}{An optional integer argument controlling the behavior of the record() function from chronicler. Default is 2.}
}
\value{
A chronicler object.
}
\description{
record_ggplot takes a ggplot_expression and an optional strict argument as input and does the following steps:
\enumerate{
\item Records the ggplot_fun function with the given strict argument using the record function from chronicler
\item Passes the ggplot_expression to the recorded ggplot_fun function
\item Returns the result of the recorded ggplot_fun function
}
}
\examples{
\dontrun{
library(ggplot2)
# Unsuccessful example
x <- record_ggplot(ggplot(data = mtcars) + geom_point(aes(y = hp, x = mpgg)))
print(x)

# Successful example
z <- record_ggplot(ggplot(data = mtcars) + geom_point(aes(y = hp, x = mpg)))
print(z)
}
}
