% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronicle.R
\name{record_many}
\alias{record_many}
\title{Decorate a list of functions}
\usage{
record_many(list_funcs, strict = 2)
}
\arguments{
\item{list_funcs}{A list of function names, as strings.}

\item{strict}{The strict parameter for record. Defaults to 2.}
}
\value{
Puts a string into the systems clipboard.
}
\description{
Decorate a list of functions
}
\details{
Functions must be entered as strings of the form "function" or "package::function".
The code gets generated and copied into the clipboard. The code can then be pasted
into the text editor.
}
\examples{
\dontrun{
list_funcs <- list("exp", "dplyr::select", "exp")
record_many(list_funcs)
}
}
